# server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bansuri Creations API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "components": {
    "securitySchemes": {
      "adminSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "admin_session"
      }
    },
...
  },
  "paths": {
...
    "/api/orders": {
      "post": {
        "operationId": "createOrder",
        "tags": [
          "Orders"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "customerName": {
                    "type": "string",
                    "minLength": 1
                  },
                  "customerPhone": {
                    "type": "string",
                    "minLength": 6
                  },
                  "customerEmail": {
                    "type": "string",
                    "format": "email"
                  },
                  "addressStreet": {
                    "type": "string",
                    "minLength": 1
                  },
                  "addressCity": {
                    "type": "string",
                    "minLength": 1
                  },
                  "addressState": {
                    "type": "string",
                    "minLength": 1
                  },
                  "addressPincode": {
                    "type": "string",
                    "minLength": 1
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "variantId": {
                          "type": "string"
                        },
                        "quantity": {
                          "type": "integer",
                          "exclusiveMinimum": 0
                        }
                      },
                      "required": [
                        "variantId",
                        "quantity"
                      ]
                    },
                    "minItems": 1
                  }
                },
                "required": [
                  "customerName",
                  "customerPhone",
                  "customerEmail",
                  "addressStreet",
                  "addressCity",
                  "addressState",
                  "addressPincode",
                  "items"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created and a Razorpay order opened for payment",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "orderId": {
                      "type": "string"
                    },
                    "orderNumber": {
                      "type": "string"
                    },
                    "razorpayOrderId": {
                      "type": "string"
                    },
                    "razorpayKeyId": {
                      "type": "string"
                    },
                    "amount": {
                      "type": "number"
                    }
                  },
                  "required": [
                    "orderId",
                    "orderNumber",
                    "razorpayOrderId",
                    "amount"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Invalid order payload",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "409": {
            "description": "Stock validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{orderNumber}": {
      "get": {
        "operationId": "getOrder",
        "tags": [
          "Orders"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": true,
            "name": "orderNumber",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "Order detail",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "orderNumber": {
                      "type": "string"
                    },
                    "customerName": {
                      "type": "string"
                    },
                    "customerPhone": {
                      "type": "string"
                    },
                    "customerEmail": {
                      "type": "string"
                    },
                    "addressStreet": {
                      "type": "string"
                    },
                    "addressCity": {
                      "type": "string"
                    },
                    "addressState": {
                      "type": "string"
                    },
                    "addressPincode": {
                      "type": "string"
                    },
                    "subtotal": {
                      "type": "number"
                    },
                    "shippingFee": {
                      "type": "number"
                    },
                    "total": {
                      "type": "number"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "PENDING",
                        "PAID",
                        "PROCESSING",
                        "SHIPPED",
                        "DELIVERED",
                        "CANCELLED"
                      ]
                    },
                    "razorpayOrderId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "razorpayPaymentId": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "paidAt": {
                      "type": [
                        "string",
                        "null"
                      ]
                    },
                    "createdAt": {
                      "type": "string"
                    },
                    "updatedAt": {
                      "type": "string"
                    },
                    "items": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "orderId": {
                            "type": "string"
                          },
                          "productVariantId": {
                            "type": "string"
                          },
                          "productNameSnapshot": {
                            "type": "string"
                          },
                          "variantLabelSnapshot": {
                            "type": "string"
                          },
                          "unitPrice": {
                            "type": "number"
                          },
                          "quantity": {
                            "type": "number"
                          }
                        },
                        "required": [
                          "id",
                          "orderId",
                          "productVariantId",
                          "productNameSnapshot",
                          "variantLabelSnapshot",
                          "unitPrice",
                          "quantity"
                        ]
                      }
                    }
                  },
                  "required": [
                    "id",
                    "orderNumber",
                    "customerName",
                    "customerPhone",
                    "customerEmail",
                    "addressStreet",
                    "addressCity",
                    "addressState",
                    "addressPincode",
                    "subtotal",
                    "shippingFee",
                    "total",
                    "status",
                    "razorpayOrderId",
                    "razorpayPaymentId",
                    "paidAt",
                    "createdAt",
                    "updatedAt",
                    "items"
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/orders": {
      "get": {
        "operationId": "adminListOrders",
        "tags": [
          "Admin Orders"
        ],
        "security": [
          {
            "adminSession": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "status",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Orders, optionally filtered by status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "orderNumber": {
                        "type": "string"
                      },
                      "customerName": {
                        "type": "string"
                      },
                      "customerPhone": {
                        "type": "string"
                      },
                      "customerEmail": {
                        "type": "string"
                      },
                      "addressStreet": {
                        "type": "string"
                      },
                      "addressCity": {
                        "type": "string"
                      },
                      "addressState": {
                        "type": "string"
                      },
                      "addressPincode": {
                        "type": "string"
                      },
                      "subtotal": {
                        "type": "number"
                      },
                      "shippingFee": {
                        "type": "number"
                      },
                      "total": {
                        "type": "number"
                      },
                      "status": {
                        "type": "string",
                        "enum": [
                          "PENDING",
                          "PAID",
                          "PROCESSING",
                          "SHIPPED",
                          "DELIVERED",
                          "CANCELLED"
                        ]
                      },
                      "razorpayOrderId": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "razorpayPaymentId": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "paidAt": {
                        "type": [
                          "string",
                          "null"
                        ]
                      },
                      "createdAt": {
                        "type": "string"
                      },
                      "updatedAt": {
                        "type": "string"
                      },
                      "items": {
                        "type": "array",
                        "items": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "orderId": {
                              "type": "string"
                            },
                            "productVariantId": {
                              "type": "string"
                            },
                            "productNameSnapshot": {
                              "type": "string"
                            },
                            "variantLabelSnapshot": {
                              "type": "string"
                            },
                            "unitPrice": {
                              "type": "number"
                            },
                            "quantity": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "id",
                            "orderId",
                            "productVariantId",
                            "productNameSnapshot",
                            "variantLabelSnapshot",
                            "unitPrice",
                            "quantity"
                          ]
                        }
                      }
                    },
                    "required": [
                      "id",
                      "orderNumber",
                      "customerName",
                      "customerPhone",
                      "customerEmail",
                      "addressStreet",
                      "addressCity",
                      "addressState",
                      "addressPincode",
                      "subtotal",
                      "shippingFee",
                      "total",
                      "status",
                      "razorpayOrderId",
                      "razorpayPaymentId",
                      "paidAt",
                      "createdAt",
                      "updatedAt",
                      "items"
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "details": {}
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}